1) added Channel to handle every fd(listenFd & connFd); 2) use TcpServer(inherited from IChannelCallBack) to be the callback of EPOLLIN; 3) use unordered_map channels_ in TcpServer to store every fd's Channel and prevented memory leak.

+++ 05/echosrv_epoll.cpp
@@ -120,7 +120,7 @@ int main() {
             ret = write(conn_fd, buf, strlen(buf));
             if (ret == -1)
                 ERR_EXIT("write");
-            if (ret < strlen(buf))
+            if (ret < static_cast<int>(strlen(buf)))
                 ERR_EXIT("write not finish one time");
         }
     }
